+++ multi/integration/networkManager/utils.lua
@@ -1,0 +1,4 @@
+local bin = require("bin")
+local utils = {}
+-- Will contain data that handles sterilizing and managing data
+return utils